Rename starknetdata.gateway package to feeder
