+++ starknetdata/feeder/gateway.go
@@ -1,4 +1,4 @@
-package gateway
+package feeder
 
 import (
 	"context"
