+++ starknetdata/feeder/gateway_pkg_test.go
@@ -1,4 +1,4 @@
-package gateway
+package feeder
 
 import (
 	"encoding/json"
